requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
ale-py==0.11.1
alembic==1.16.4
annotated-types==0.7.0
attrs==25.3.0
casadi==3.7.0
certifi==2025.7.14
cfgv==3.4.0
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
copier==9.7.1
cycler==0.12.1
datasets==4.0.0
DateTime==5.5
dill==0.3.8
distlib==0.3.9
docopt-ng==0.9.0
dunamai==1.24.1
etils==1.12.2
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.9.0
grpcio==1.73.0
gymnasium==1.2.0
-e git+https://github.com/derekc22/UR3e.git@7443c318780761cc7dcd117e800a533156513cca#egg=gymnasium_env
hf-xet==1.1.5
huggingface-hub==0.34.1
huggingface-sb3==3.0
identify==2.6.12
idna==3.10
imageio==2.37.0
imitation==1.0.1
importlib_resources==6.5.2
Jinja2==3.1.6
jinja2-ansible-filters==1.3.2
joblib==1.5.1
jsonpickle==4.1.1
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
mujoco==3.3.3
multidict==6.6.3
multiprocess==0.70.16
munch==4.0.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.0
opencv-python==4.11.0.86
optuna==4.4.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
plumbum==1.9.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.6
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
questionary==2.1.0
requests==2.32.4
rich==14.0.0
sacred==0.8.7
scikit-learn==1.7.1
scipy==1.15.3
seals==0.2.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
stable_baselines3==2.7.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
wasabi==1.1.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
